refactor: centralize hook logic
