--- packages/serverless-cms-aws/src/admin/plugins/index.ts
@@ -1,1 +1,0 @@
-export * from "./ensureApiDeployed";
