+++ packages/serverless-cms-aws/src/website/plugins/index.ts
@@ -2,4 +2,3 @@ export * from "./renderWebsite";
 export * from "./generateCommonHandlers";
 export * from "./lambdaEdgeWarning";
 export * from "./telemetryNoLongerNewUser";
-export * from "./ensureApiDeployed";
